drivers: boards: Share board_crashdump() implementation
