+++ src/drivers/boards/aerofc-v1/CMakeLists.txt
@@ -35,6 +35,7 @@ px4_add_module(
 	MODULE drivers__boards__aerofc-v1
 	COMPILE_FLAGS
 	SRCS
+		../common/stm32/board_crashdump.c
 		aerofc_init.c
 		aerofc_spi.c
 		aerofc_usb.c
