+++ src/drivers/boards/crazyflie/CMakeLists.txt
@@ -33,6 +33,7 @@
 px4_add_module(
 	MODULE drivers__boards__crazyflie
 	SRCS
+		../common/stm32/board_crashdump.c
 		../common/board_dma_alloc.c
 		crazyflie_init.c
 		crazyflie_usb.c
