+++ src/drivers/boards/px4fmu-v4/CMakeLists.txt
@@ -34,6 +34,7 @@ px4_add_module(
 	MODULE drivers__boards__px4fmu-v4
 	COMPILE_FLAGS
 	SRCS
+		../common/stm32/board_crashdump.c
 		../common/board_dma_alloc.c
 		px4fmu_can.c
 		px4fmu_init.c
