+++ src/drivers/boards/px4nucleoF767ZI-v1/CMakeLists.txt
@@ -35,6 +35,7 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Os
 	SRCS
+		../common/stm32/board_crashdump.c
 		../common/board_dma_alloc.c
 #		WIP px4nucleo_can.c
 		px4nucleo_init.c
