+++ src/drivers/boards/tap-v1/CMakeLists.txt
@@ -35,6 +35,7 @@ px4_add_module(
 	MODULE drivers__boards__tap-v1
 	COMPILE_FLAGS
 	SRCS
+		../common/stm32/board_crashdump.c
 		../common/board_dma_alloc.c
 		tap_init.c
 		tap_pwr.c
